rtp: move rewriter into its own `packet` package
