+++ pkg/packet/rtp.go
@@ -1,4 +1,4 @@
-package subscription
+package packet
 
 import (
 	"github.com/pion/rtp"
